mysql: move `datetime` from `evalengine`
