+++ go/mysql/datetime/spec.go
