+++ go/vt/vtgate/evalengine/compiler_fn.go
@@ -18,10 +18,10 @@ package evalengine
 
 import (
 	"vitess.io/vitess/go/mysql/collations"
+	"vitess.io/vitess/go/mysql/datetime"
 	"vitess.io/vitess/go/sqltypes"
 	"vitess.io/vitess/go/vt/proto/vtrpc"
 	"vitess.io/vitess/go/vt/vterrors"
-	"vitess.io/vitess/go/vt/vtgate/evalengine/internal/datetime"
 )
 
 func (c *compiler) compileFn(call callable) (ctype, error) {
