+++ go/vt/vtgate/evalengine/fn_time.go
@@ -19,9 +19,9 @@ package evalengine
 import (
 	"time"
 
+	"vitess.io/vitess/go/mysql/datetime"
 	"vitess.io/vitess/go/sqltypes"
 	querypb "vitess.io/vitess/go/vt/proto/query"
-	"vitess.io/vitess/go/vt/vtgate/evalengine/internal/datetime"
 )
 
 var SystemTime = time.Now
